separating ground functionalities from implementation that require on disk
